\name{Spectra file}
\Rdversion{1.1}
\title{Example data sets}
\description{This data table contains NIR spectra recorded at World Agroforestry Centre- ICRAF using Bruker's MPA spectrometer. The samples were collected from different parts of the world. 
Another table with site details will be made available.
}

\author{Andrew Sila}

\examples{
# To load the file use; require(soil.spec) nir<-load(Soil NIR)}
