\name{soil.spec-package}
\alias{soil.spec-package}
\alias{soil.spec}
\docType{package}
\title{Soil spectral data exploration and regression functions}
\description{This package combines existing R functions with new code for 	soil spectral analysis. The result is an easy to use tool for (i) 	importing of spc-files, (ii) principal component analysis, (iii) 	sample selection using the Kennard-Stone algorithm, (iv) spectral 	transformation and (v) comparison of regression methods.}
\details{
\tabular{ll}{
Package: \tab soil.spec\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2010-07-10\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Thomas Terhoeven-Urselmans

Maintainer: Thomas Terhoeven-Urselmans <t.urselmans@cgiar.org>
}
