\name{soil.spec-package}
\alias{soil.spec-package}
\alias{soil.spec}
\docType{package}
\title{Soil spectral data exploration and regression functions}
\description{This package combines existing R functions for table manipulation, file transfer and data summaries with new code for 	soil spectral analysis. The result is an easy to use tool for (i) 	importing of spc-files, (ii) principal component analysis, (iii) 	sample selection using the Kennard-Stone algorithm, (iv) spectral 	transformation and (v) comparison of regression methods. In our next update of tools for outlier detection, functionality for randomForest and MCMC models will be added together with a vignette which will contain Africa soil spectral data}
\details{
\tabular{ll}{
Package: \tab soil.spec\cr
Type: \tab Package\cr
Version: \tab 2.1.1\cr
Date: \tab 2013-07-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andrew Sila and Thomas Terhoeven-Urselmans

Maintainer: Andrew Sila <a.sila@cgiar.org>
}

\keyword{Spectroscopy}
\keyword{Chemometrics}
\keyword{MIR}
\keyword{NIR}
\keyword{Calibration}